Move ReferenceData
Move to strata-basics root package
