+++ modules/basics/src/test/java/com/opengamma/strata/basics/ReferenceDataTest.java
@@ -19,7 +19,6 @@ import java.util.Optional;
 import org.testng.annotations.Test;
 
 import com.google.common.collect.ImmutableMap;
-import com.opengamma.strata.basics.market.ImmutableReferenceData;
 
 /**
  * Test {@link ReferenceData} and {@link ImmutableReferenceData}.
