+++ modules/basics/src/test/java/com/opengamma/strata/basics/date/HolidayCalendarIdTest.java
@@ -15,9 +15,9 @@ import static org.testng.Assert.assertSame;
 import org.testng.annotations.Test;
 
 import com.google.common.collect.ImmutableMap;
+import com.opengamma.strata.basics.ImmutableReferenceData;
 import com.opengamma.strata.basics.ReferenceData;
 import com.opengamma.strata.basics.ReferenceDataNotFoundException;
-import com.opengamma.strata.basics.market.ImmutableReferenceData;
 
 /**
  * Test {@link HolidayCalendarId}.
