+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/swap/e2e/SwapPricePerformance.java
@@ -22,10 +22,10 @@ import static com.opengamma.strata.product.common.PayReceive.RECEIVE;
 
 import java.time.LocalDate;
 
+import com.opengamma.strata.basics.ImmutableReferenceData;
 import com.opengamma.strata.basics.ReferenceData;
 import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.basics.date.DaysAdjustment;
-import com.opengamma.strata.basics.market.ImmutableReferenceData;
 import com.opengamma.strata.basics.schedule.Frequency;
 import com.opengamma.strata.basics.schedule.PeriodicSchedule;
 import com.opengamma.strata.basics.schedule.StubConvention;
