+++ modules/product/src/test/java/com/opengamma/strata/product/SecurityTradeTest.java
@@ -13,7 +13,7 @@ import static org.testng.Assert.assertEquals;
 
 import org.testng.annotations.Test;
 
-import com.opengamma.strata.basics.market.ImmutableReferenceData;
+import com.opengamma.strata.basics.ImmutableReferenceData;
 
 /**
  * Test {@link SecurityTrade}.
